created models.py to better organize the models
